% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_DE_analysis.R
\name{convert_ID}
\alias{convert_ID}
\title{Convert IDs using biomaRt human ensembl database}
\usage{
convert_ID(gene_list, from, to, description = FALSE)
}
\arguments{
\item{gene_list}{A vector containing the terms to be converted}

\item{from}{The format of the input data}

\item{to}{The format of the desired output}

\item{description}{If changed to TRUE will include an addtional column with description of term}
}
\value{
description
}
\description{
Convert IDs using biomaRt human ensembl database
}
\examples{
ID_converted <- convert_ID(gene_list, from = "ensembl_gene_id", to = "external_gene_name")
}
