% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_query.R
\name{get_subtypes}
\alias{get_subtypes}
\title{Show subtypes for a TCGA project}
\usage{
get_subtypes(project)
}
\arguments{
\item{project}{Name of TCGA project}
}
\value{
List of subtypes associated with TCGA project
}
\description{
Show subtypes for a TCGA project
}
\examples{
subtypes <- get_subtypes(BRCA)
}
