% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_query.R
\name{get_TCGA_RNAseq_data}
\alias{get_TCGA_RNAseq_data}
\title{Retreive transcriptomic data from TCGA}
\usage{
get_TCGA_RNAseq_data(
  project_name,
  subtype,
  view_clinical = FALSE,
  manifest_file = FALSE
)
}
\arguments{
\item{project_name}{The TCGA project name}

\item{subtype}{The cancer subtype}

\item{manifest_file}{Opt to keep manifest file that comes with downloaded GDC data}

\item{view_clincal}{Boolean value to view clinical figo stage of samples}
}
\value{
Data frame with unstranded counts
}
\description{
Retreive transcriptomic data from TCGA
}
\examples{
data <- get_TCGA_RNAseq_data(TCGA-BRCA, LumB)
}
