% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_DE_analysis.R
\name{remove_low_activity_genes}
\alias{remove_low_activity_genes}
\title{Removes low activity transcripts from data. Requires a file containing all samples including healthy/control and disease groups.}
\usage{
remove_low_activity_genes(data, top_x_samples, min_samples)
}
\arguments{
\item{data}{data contianing healthy/control and disease groups}

\item{top_x_samples}{percentile range that a trasncript must appear in}

\item{min_sample}{Percentage value indicating the number of samples the trasncript must appear within the range specified in `top_x_samples`}
}
\value{
data frame with low activity genes removed
}
\description{
Removes low activity transcripts from data. Requires a file containing all samples including healthy/control and disease groups.
}
\examples{
treated_data <- remove_low_activity_genes(data, top_x_samples = 0.50, min_samples = 0.10)
}
