% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_DE_analysis.R
\name{TCGA_DE_analysis}
\alias{TCGA_DE_analysis}
\title{Performs differential expression analysis on TCGA data}
\usage{
TCGA_DE_analysis(data, show_plots = TRUE)
}
\arguments{
\item{data}{Raw counts from TCGA and healthy cell group}

\item{show_plots}{Displays BCV, smear and Pvalu plots}
}
\value{
Dataframe of significantly differentially expressed genes as well as plots.
}
\description{
Performs differential expression analysis on TCGA data
}
\examples{
# example code
DE_results <- TCGA_DE_analysis(data)
}
