% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_DE_analysis.R
\name{TCGA_GTEx_combine}
\alias{TCGA_GTEx_combine}
\title{combines TCGA and GTEx data for preprocessing}
\usage{
TCGA_GTEx_combine(TCGA_data, GTEx_data)
}
\arguments{
\item{TCGA_data}{TCGA unstranded data. Output by `get_TCGA_RNAseq_data`}

\item{GTEx_data}{Data downloaded from the GTEx portal}
}
\value{
Returns a dataframe with the two input datas combined, ready for differential expression analysis
}
\description{
combines TCGA and GTEx data for preprocessing
}
\examples{
GTEx_data <- read.table("bulk-gex_v8_rna-seq_counts-by-tissue_gene_reads_2017-06-05_v8_breast_mammary_tissue.gct", skip = 2, header = T)
data <- TCGA_GTEx_combine(TCGA_data, GTEx_data)
}
